[fix] structure for cleaner method call
